requirements.txt
openai==0.28
python-dotenv
numpy
transformers
torch
fastapi
uvicorn